# contracts/openapi/forge-network.openapi.yaml
openapi: 3.0.3
info:
  title: Forge Network API
  version: 0.1.0
  description: |
    Provider-independent cluster address plan and lease API (epic 22 / step 22.01).
    Owns cluster CIDR → per-node blocks → per-workload addresses.
servers:
  - url: http://127.0.0.1:4110
    description: Local Compose (host port 4110)
paths:
...
  /v1/networks/{name}/workload-leases:
    parameters:
      - $ref: "#/components/parameters/NetworkName"
    get:
      operationId: listWorkloadLeases
      summary: List active workload overlay address leases
      responses:
        "200":
          description: Active leases
          content:
            application/json:
              schema:
                type: object
                properties:
                  leases:
                    type: array
                    items:
                      type: object
                      properties:
                        workload_id:
                          type: string
                        node_id:
                          type: string
                        address:
                          type: string
              example:
                leases:
                  - workload_id: wl_123
                    node_id: node-a
                    address: "10.100.1.5"
        "404":
          description: Network not found
    post:
      operationId: allocateWorkloadLease
      summary: Allocate a workload address inside a node block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocateWorkloadLeaseRequest"
            example:
              node_id: node-a
              workload_id: wl_123
      responses:
        "200":
          description: Workload lease (idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkloadLease"
              example:
                workload_id: wl_123
                address: "10.100.1.5"
        "409":
          description: Node block exhausted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                NodeBlockExhausted:
                  value:
                    error:
                      code: NodeBlockExhausted
                      message: node block has no free addresses
                      requestId: req_example
  /v1/networks/{name}/route-drift:
    parameters:
      - $ref: "#/components/parameters/NetworkName"
    post:
      operationId: reportRouteDrift
      summary: Report Discovery/Network/route drift observations (22.06)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drift_count]
              properties:
                drift_count:
                  type: integer
                  format: int64
            example:
              drift_count: 1
      responses:
        "202":
          description: Recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  drift_count:
                    type: integer
                    format: int64
  /v1/networks/{name}/workload-leases/{id}:
    parameters:
      - $ref: "#/components/parameters/NetworkName"
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: workload_id
    delete:
      operationId: releaseWorkloadLease
      summary: Release a workload address
      responses:
        "204":
          description: Released
        "404":
          description: Not found
...
  /v1/nodes/{node_id}/network-policy-rules:
    parameters:
      - $ref: "#/components/parameters/NodeId"
    get:
      operationId: getNetworkPolicyRules
      summary: Fetch the compiled NetworkPolicy rule set for a node
      responses:
        "200":
          description: Per-node rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkPolicyRules"
              example:
                node_id: node-a
                generation: 7
                rules:
                  - workload_id: wl_1
                    direction: ingress
                    from_cidr: "10.100.2.5/32"
                    port: 8080
                    protocol: tcp
                    action: allow
                  - workload_id: wl_1
                    direction: egress
                    to_cidr: "0.0.0.0/0"
                    action: deny
                    reason: default-deny-environment
  /v1/nodes/{node_id}/network-policy-denied:
    parameters:
      - $ref: "#/components/parameters/NodeId"
    post:
      operationId: reportNetworkPolicyDenied
      summary: Report a denied connection from Runtime enforcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_workload, to_workload, port]
              properties:
                from_workload:
                  type: string
                to_workload:
                  type: string
                port:
                  type: integer
                protocol:
                  type: string
                reason:
                  type: string
      responses:
        "202":
          description: Recorded (metric + network.policy.denied event)
  /v1/workload-placements/{workload_id}:
    parameters:
      - name: workload_id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: upsertWorkloadPlacement
      summary: Upsert a scheduler placement mirror used by PolicyCompiler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkloadPlacement"
      responses:
        "200":
          description: Upserted
...
components:
  parameters:
    NetworkName:
      name: name
      in: path
      required: true
      schema:
        type: string
      example: cluster-overlay
    NodeId:
      name: node_id
      in: path
      required: true
      schema:
        type: string
      example: node-a
  schemas:
...
    AllocateWorkloadLeaseRequest:
      type: object
      required: [node_id, workload_id]
      properties:
        node_id:
          type: string
        workload_id:
          type: string
          example: wl_123
    WorkloadLease:
      type: object
      required: [workload_id, address]
      properties:
        workload_id:
          type: string
        address:
          type: string
          example: "10.100.1.5"
...
    NetworkPolicyRules:
      type: object
      required: [node_id, generation, rules]
      properties:
        node_id:
          type: string
          example: node-a
        generation:
          type: integer
          format: int64
          example: 7
        rules:
          type: array
          items:
            $ref: "#/components/schemas/NetworkPolicyRule"
    NetworkPolicyRule:
      type: object
      required: [workload_id, direction, action]
      properties:
        workload_id:
          type: string
        direction:
          type: string
          enum: [ingress, egress]
        from_cidr:
          type: string
        to_cidr:
          type: string
        port:
          type: integer
        protocol:
          type: string
        action:
          type: string
          enum: [allow, deny]
        reason:
          type: string
          example: default-deny-environment
    WorkloadPlacement:
      type: object
      required: [organization, project, environment, node_id]
      properties:
        workload_id:
          type: string
        organization:
          type: string
        project:
          type: string
        environment:
          type: string
        node_id:
          type: string
        application:
          type: string
        service:
          type: string
        database:
          type: string
        queue:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - CidrCollision
                - NoAddressSpaceAvailable
                - NodeBlockExhausted
                - NetworkNotReady
                - not_found
                - invalid_request
                - already_exists
                - internal_error
            message:
              type: string
            requestId:
              type: string
            details:
              type: object
              additionalProperties:
                type: string